chatRoomName이 공백만으로 설정되지 않도록 예외 처리
