+++ chat-service/src/main/java/com/ovengers/chatservice/mysql/exception/InvalidChatRoomNameException.java
@@ -1,0 +1,7 @@
+package com.ovengers.chatservice.mysql.exception;
+
+public class InvalidChatRoomNameException extends RuntimeException {
+    public InvalidChatRoomNameException(String message) {
+        super(message);
+    }
+}
