+++ chat-service/src/main/java/com/ovengers/chatservice/mysql/service/ChatRoomService.java
@@ -2,7 +2,9 @@ package com.ovengers.chatservice.mysql.service;
 
 import com.ovengers.chatservice.mysql.dto.ChatRoomDto;
 import com.ovengers.chatservice.mysql.entity.ChatRoom;
+import com.ovengers.chatservice.mysql.exception.InvalidChatRoomNameException;
 import com.ovengers.chatservice.mysql.repository.ChatRoomRepository;
+import io.micrometer.common.util.StringUtils;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.Getter;
@@ -31,8 +33,9 @@ public class ChatRoomService {
 
     // 채팅방 생성
     public ChatRoomDto createChatRoom(String name) {
+        validateChatRoomName(name); // 유효성 검사 추가
         ChatRoom chatRoom = ChatRoom.builder()
-                .name(name)
+                .name(name.trim()) // 이름 양끝 공백 제거
                 .createdAt(LocalDateTime.now())
                 .build();
         return ChatRoomDto.fromEntity(chatRoomRepository.save(chatRoom));
@@ -40,10 +43,18 @@ public class ChatRoomService {
 
     // 채팅방 이름 수정
     public ChatRoomDto updateChatRoom(Long chatRoomId, String newName) {
+        validateChatRoomName(newName); // 유효성 검사 추가
         ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId)
-                .orElseThrow(() -> new EntityNotFoundException("ChatRoom not found with id: " + chatRoomId));
-        chatRoom.setName(newName);
+                .orElseThrow(() -> new EntityNotFoundException(chatRoomId + "번 채팅방은 존재하지 않습니다."));
+        chatRoom.setName(newName.trim()); // 이름 양끝 공백 제거
         return ChatRoomDto.fromEntity(chatRoomRepository.save(chatRoom));
+    }
+
+    // 유효성 검사 메서드
+    private void validateChatRoomName(String name) {
+        if (StringUtils.isBlank(name)) { // Apache Commons Lang 사용 (공백 또는 null 확인)
+            throw new InvalidChatRoomNameException("채팅방 이름은 공백만으로 지정할 수 없습니다.");
+        }
     }
 
     // 채팅방 삭제
